+++ plugin/app-chart/data-management.css
@@ -1,0 +1,41 @@
+.app-chart-data-management {
+    max-width: 600px;
+}
+.data-stats {
+    margin-bottom: 20px;
+    background-color: rgba(255, 255, 255, 0.5);
+    border-radius: 8px;
+    padding: 15px;
+}
+.data-actions {
+    display: flex;
+    gap: 10px;
+    margin-bottom: 15px;
+}
+.data-message {
+    margin-top: 15px;
+    padding: 10px;
+    border-radius: 4px;
+}
+.data-message.success {
+    background-color: rgba(46, 160, 67, 0.1);
+    border: 1px solid rgba(46, 160, 67, 0.2);
+    color: #2ea043;
+}
+.data-message.error {
+    background-color: rgba(248, 81, 73, 0.1);
+    border: 1px solid rgba(248, 81, 73, 0.2);
+    color: #f85149;
+}
+.data-stat-item {
+    display: flex;
+    justify-content: space-between;
+    padding: 8px 0;
+    border-bottom: 1px solid rgba(128, 128, 128, 0.2);
+}
+.data-stat-item:last-child {
+    border-bottom: none;
+}
+.data-stat-label {
+    font-weight: bold;
+}
