+++ plugin/app-chart/data-management.html
@@ -1,0 +1,14 @@
+<div class="app-chart-data-management">
+    <div class="data-stats">
+        <div class="loading">加载数据统计中...</div>
+    </div>
+
+    <div class="data-actions">
+        <button id="refresh-stats" class="btn btn-secondary">刷新统计</button>
+        <button id="clear-data" class="btn btn-danger">清除所有数据</button>
+        <button id="export-data" class="btn btn-primary">导出数据</button>
+        <button id="download-db" class="btn btn-primary">下载数据库</button>
+    </div>
+
+    <div id="data-message" class="data-message" style="display: none;"></div>
+</div>
