+++ plugin/app-chart/settings.css
@@ -1,0 +1,24 @@
+.app-chart-settings {
+    max-width: 600px;
+}
+.app-chart-settings small {
+    display: block;
+    color: #666;
+    margin-top: 2px;
+    font-size: 12px;
+}
+.settings-message {
+    margin-top: 15px;
+    padding: 10px;
+    border-radius: 4px;
+}
+.settings-message.success {
+    background-color: rgba(46, 160, 67, 0.1);
+    border: 1px solid rgba(46, 160, 67, 0.2);
+    color: #2ea043;
+}
+.settings-message.error {
+    background-color: rgba(248, 81, 73, 0.1);
+    border: 1px solid rgba(248, 81, 73, 0.2);
+    color: #f85149;
+}
