+++ plugin/app-chart/settings.html
@@ -1,0 +1,36 @@
+<div class="app-chart-settings">
+    <div class="form-group">
+        <label for="db_path">数据库路径:</label>
+        <input type="text" id="db_path" value="{{ db_path }}" />
+        <small>数据库文件的路径，相对于应用根目录</small>
+    </div>
+    <div class="form-group">
+        <label for="max_apps">最大应用数量:</label>
+        <input type="number" id="max_apps" value="{{ max_apps }}" min="1" max="50" />
+        <small>图表中显示的最大应用数量</small>
+    </div>
+    <div class="form-group">
+        <label for="default_days">默认显示天数:</label>
+        <select id="default_days">
+            <option value="7" {% if default_days == 7 %}selected{% endif %}>7天</option>
+            <option value="14" {% if default_days == 14 %}selected{% endif %}>14天</option>
+            <option value="30" {% if default_days == 30 %}selected{% endif %}>30天</option>
+        </select>
+        <small>默认显示的时间范围</small>
+    </div>
+    <div class="form-group">
+        <label for="chart_height">图表高度:</label>
+        <input type="number" id="chart_height" value="{{ chart_height }}" min="100" max="800" step="10" />
+        <small>图表的高度（像素）</small>
+    </div>
+    <div class="form-group">
+        <label for="colors">图表颜色:</label>
+        <textarea id="colors" rows="5">{{ colors_str }}</textarea>
+        <small>每行一个颜色代码，例如 #4e79a7</small>
+    </div>
+    <div class="form-actions">
+        <button id="save-settings" class="btn btn-primary">保存设置</button>
+        <button id="reset-settings" class="btn btn-secondary">重置</button>
+    </div>
+    <div id="settings-message" class="settings-message" style="display: none;"></div>
+</div>
